Add interactive functionality to RecidivizForm component
Merge test fixes
