+++ src/components/App/App.test.js
@@ -3,6 +3,8 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 jest.mock('../../lib/static/recidivismSlider.js');
+jest.mock('../../lib/static/nlform.js');
+jest.mock('../../lib/static/likelihoodSelect.js');
 
 it('renders the App component', () => {
   const div = document.createElement('div');
